[feature] 게시글 없을 경우 예외 처리

+++ src/main/java/com/resume/service/NoticeService.java
@@ -3,13 +3,13 @@ package com.resume.service;
 import com.resume.Repository.NoticeDao;
 import com.resume.dto.NoticeDto;
 import com.resume.dto.SearchCondition;
+import com.resume.exception.PostNotFountException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Map;
 
 @Service
 @Slf4j
@@ -27,24 +27,27 @@ public class NoticeService {
     }
 
     public int deleteNotice(int num){
+        NoticeDto findNotice = dao.findByNum(num);
+        if(findNotice == null){
+            throw new PostNotFountException("존재하지 않는 게시글입니다.");
+        }
         return dao.deleteNotice(num);
     }
     public NoticeDto findByNum(int num){
+        NoticeDto findNotice = dao.findByNum(num);
+        if(findNotice == null){
+            throw new PostNotFountException("존재하지 않는 게시글입니다.");
+        }
         return dao.findByNum(num);
     }
 
-    public List<NoticeDto> findAll(Map map){
-        return dao.findAll(map);
-    }
 
     public int searchResultCnt(SearchCondition sc){
         return dao.searchResultCnt(sc);
     }
-//    @Cacheable("postComments")
-
+    @Cacheable("postComments")
     public List<NoticeDto>searchSelectPage(SearchCondition sc){
         List<NoticeDto> result = dao.searchSelectPage(sc);
-        log.info("searchSelectPage 호출");
         return result;
     }
